refactor: extract pending-message helpers out of page.tsx
The OAuth-redirect pending-message persistence (storage key, type, and
save/load/has helpers) lived inline at the top of the 1600-line page.tsx.
Move it into a focused lib/pending-message.ts module and replace the raw
sessionStorage key check in page.tsx with a hasPendingMessage() helper so
the storage key no longer leaks into the component.
